+++ soroban-sdk/src/env.rs
@@ -49,7 +49,7 @@ pub use internal::Val;
 pub type EnvVal = internal::EnvVal<Env, RawVal>;
 pub type EnvObj = internal::EnvVal<Env, Object>;
 
-use crate::{deploy::Deployer, logger::Logger, Bytes, BytesN, ContractData, Events, Ledger, Vec};
+use crate::{deploy::Deployer, logging::Logger, Bytes, BytesN, ContractData, Events, Ledger, Vec};
 
 /// The [Env] type provides access to the environment the contract is executing
 /// within.
@@ -132,7 +132,7 @@ impl Env {
 
     /// Get the [Logger] for logging debug events.
     #[inline(always)]
-    pub fn log(&self) -> Logger {
+    pub fn logger(&self) -> Logger {
         Logger::new(self)
     }
